Low level programming

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	int i;
-	float f[98];
+	double f[98];
 
 	f[0] = 1;
 	f[1] = 2;
